impove ordering mixin
